new changes

+++ src/com/interview/array/ArrayRearrange.java
@@ -10,10 +10,10 @@ package com.interview.array;
 
 public class ArrayRearrange {
 
-    public static void rearrange(int arr[], int n) {
+    public static void rearrange(long arr[], int n) {
         // First step: Increase all values by (arr[arr[i]]%n)*n
         for (int i = 0; i < n; i++)
-            arr[i] = arr[i] + (arr[arr[i]] % n) * n;
+            arr[i] = arr[i] + (arr[(int)arr[i]] % n) * n;
 
         // Second Step: Divide all values by n
         for (int i = 0; i < n; i++)
@@ -21,7 +21,7 @@ public class ArrayRearrange {
     }
 
     public static void main(String[] args) {
-        int[] arr  = {3, 2, 0, 1};
+        long[] arr  = {3, 2, 0, 1};
         rearrange(arr, 4);
 
         for(int i=0; i<arr.length; i++){
